+++ src/hx20-devices/disk/tf20drivediskimage.hpp
@@ -29,13 +29,13 @@ public:
                              uint8_t new_us, uint8_t const *new_filename, uint8_t new_extent) override;
     //0x21
     virtual uint8_t file_read(void *fcb, uint32_t record,
-                              uint8_t &cur_extent, uint16_t &cur_record, void *buffer) override;
+                              uint8_t &cur_extent, uint8_t &cur_record, void *buffer) override;
     //0x22
     virtual uint8_t file_write(void *fcb, void const *buffer, uint32_t record,
-                               uint8_t &cur_extent, uint16_t &cur_record) override;
+                               uint8_t &cur_extent, uint8_t &cur_record) override;
     //0x23
     virtual uint8_t file_size(void *fcb, uint8_t &extent,
-                              uint16_t &record, uint32_t &records) override;
+                              uint8_t &record, uint32_t &records) override;
     //0x24
     virtual uint8_t file_tell(void *fcb, uint32_t &records) override;
     //0x7b
